refactor: 파일 분리로 가독성 확보
